Extract app server process and queue processor
