+++ src/server/codexAppServerBridge.ts
@@ -87,6 +87,15 @@ import {
   writePinnedThreadIds,
   writeThreadTitleCache,
 } from './globalStateStore.js'
+import { AppServerProcess, type StreamEventFrame } from './appServerProcess.js'
+import {
+  BackendQueueProcessor,
+  appendThreadQueuedMessage,
+  buildHeartbeatQueuedMessage,
+  normalizeThreadQueueState,
+  readThreadQueueState,
+  writeThreadQueueState,
+} from './backendQueueProcessor.js'
 import { mergeSessionSkillInputsIntoThreadResult, sanitizeThreadTurnsInlinePayloads } from './threadInlinePayloads.js'
 export { mergeSessionSkillInputsIntoTurns, sanitizeThreadTurnsInlinePayloads } from './threadInlinePayloads.js'
 
@@ -1345,243 +1354,6 @@ function getCodexAutomationsDir(): string {
   return join(getCodexHomeDir(), 'automations')
 }
 
-const THREAD_QUEUE_STATE_KEY = 'thread-queue-state'
-
-type StoredQueuedMessage = {
-  id: string
-  text: string
-  imageUrls: string[]
-  skills: Array<{ name: string; path: string }>
-  fileAttachments: Array<{ label: string; path: string; fsPath: string }>
-  collaborationMode: 'default' | 'plan'
-}
-
-type ThreadQueueState = Record<string, StoredQueuedMessage[]>
-
-type BackendQueuedTurn = {
-  threadId: string
-  message: StoredQueuedMessage
-}
-
-type ThreadQueueStateUpdate<T> = {
-  nextState: ThreadQueueState
-  result: T
-}
-
-type ResolvedCollaborationModeSettings = {
-  model: string
-  reasoningEffort: ReasoningEffort | null
-}
-
-function normalizeStoredQueuedMessage(value: unknown): StoredQueuedMessage | null {
-  const record = asRecord(value)
-  if (!record) return null
-
-  const id = typeof record.id === 'string' ? record.id.trim() : ''
-  if (!id) return null
-
-  const normalizeNamedPathItems = (items: unknown): Array<{ name: string; path: string }> => {
-    if (!Array.isArray(items)) return []
-    return items.flatMap((item) => {
-      const itemRecord = asRecord(item)
-      if (!itemRecord) return []
-      const name = typeof itemRecord.name === 'string' ? itemRecord.name.trim() : ''
-      const path = typeof itemRecord.path === 'string' ? itemRecord.path.trim() : ''
-      return name && path ? [{ name, path }] : []
-    })
-  }
-
-  const normalizeFileAttachments = (items: unknown): Array<{ label: string; path: string; fsPath: string }> => {
-    if (!Array.isArray(items)) return []
-    return items.flatMap((item) => {
-      const itemRecord = asRecord(item)
-      if (!itemRecord) return []
-      const label = typeof itemRecord.label === 'string' ? itemRecord.label.trim() : ''
-      const path = typeof itemRecord.path === 'string' ? itemRecord.path.trim() : ''
-      const fsPath = typeof itemRecord.fsPath === 'string' ? itemRecord.fsPath.trim() : ''
-      return label && path && fsPath ? [{ label, path, fsPath }] : []
-    })
-  }
-
-  return {
-    id,
-    text: typeof record.text === 'string' ? record.text : '',
-    imageUrls: normalizeStringArray(record.imageUrls),
-    skills: normalizeNamedPathItems(record.skills),
-    fileAttachments: normalizeFileAttachments(record.fileAttachments),
-    collaborationMode: record.collaborationMode === 'plan' ? 'plan' : 'default',
-  }
-}
-
-function normalizeThreadQueueState(value: unknown): ThreadQueueState {
-  const record = asRecord(value)
-  if (!record) return {}
-
-  const state: ThreadQueueState = {}
-  for (const [threadId, rawMessages] of Object.entries(record)) {
-    const normalizedThreadId = threadId.trim()
-    if (!normalizedThreadId || !Array.isArray(rawMessages)) continue
-    const messages = rawMessages.flatMap((item) => {
-      const message = normalizeStoredQueuedMessage(item)
-      return message ? [message] : []
-    })
-    if (messages.length > 0) {
-      state[normalizedThreadId] = messages
-    }
-  }
-  return state
-}
-
-let threadQueueMutationChain: Promise<unknown> = Promise.resolve()
-
-async function readThreadQueueState(): Promise<ThreadQueueState> {
-  const statePath = getCodexGlobalStatePath()
-  try {
-    const raw = await readFile(statePath, 'utf8')
-    const payload = asRecord(JSON.parse(raw)) ?? {}
-    return normalizeThreadQueueState(payload[THREAD_QUEUE_STATE_KEY])
-  } catch {
-    return {}
-  }
-}
-
-async function writeThreadQueueStateUnlocked(nextState: ThreadQueueState): Promise<void> {
-  const statePath = getCodexGlobalStatePath()
-  let payload: Record<string, unknown> = {}
-  try {
-    const raw = await readFile(statePath, 'utf8')
-    payload = asRecord(JSON.parse(raw)) ?? {}
-  } catch {
-    payload = {}
-  }
-  const normalized = normalizeThreadQueueState(nextState)
-  if (Object.keys(normalized).length > 0) {
-    payload[THREAD_QUEUE_STATE_KEY] = normalized
-  } else {
-    delete payload[THREAD_QUEUE_STATE_KEY]
-  }
-  await writeFile(statePath, JSON.stringify(payload), 'utf8')
-}
-
-async function withThreadQueueStateUpdate<T>(
-  update: (state: ThreadQueueState) => ThreadQueueStateUpdate<T> | Promise<ThreadQueueStateUpdate<T>>,
-): Promise<T> {
-  const run = threadQueueMutationChain.then(async () => {
-    const currentState = await readThreadQueueState()
-    const { nextState, result } = await update(currentState)
-    await writeThreadQueueStateUnlocked(nextState)
-    return result
-  })
-  threadQueueMutationChain = run.catch(() => {})
-  return run
-}
-
-async function writeThreadQueueState(nextState: ThreadQueueState): Promise<void> {
-  await withThreadQueueStateUpdate(() => ({
-    nextState: normalizeThreadQueueState(nextState),
-    result: undefined,
-  }))
-}
-
-async function appendThreadQueuedMessage(threadId: string, message: StoredQueuedMessage): Promise<void> {
-  const normalizedThreadId = threadId.trim()
-  if (!normalizedThreadId) throw new Error('threadId is required')
-  await withThreadQueueStateUpdate((state) => ({
-    nextState: {
-      ...state,
-      [normalizedThreadId]: [...(state[normalizedThreadId] ?? []), message],
-    },
-    result: undefined,
-  }))
-}
-
-function normalizeReasoningEffort(value: unknown): ReasoningEffort | '' {
-  const allowed: ReasoningEffort[] = ['none', 'minimal', 'low', 'medium', 'high', 'xhigh']
-  return typeof value === 'string' && allowed.includes(value as ReasoningEffort)
-    ? (value as ReasoningEffort)
-    : ''
-}
-
-function normalizeCollaborationModeReasoningEffort(value: ReasoningEffort | '' | null | undefined): ReasoningEffort | null {
-  return value && value.length > 0 ? value : null
-}
-
-function extractLocalImagePathFromUrl(value: string): string | null {
-  if (!value) return null
-  try {
-    const parsed = new URL(value, 'http://localhost')
-    if (parsed.pathname !== '/codex-local-image') return null
-    const path = parsed.searchParams.get('path')?.trim() ?? ''
-    return path.length > 0 ? path : null
-  } catch {
-    return null
-  }
-}
-
-function buildTextWithAttachments(prompt: string, files: StoredQueuedMessage['fileAttachments']): string {
-  if (files.length === 0) return prompt
-  let prefix = '# Files mentioned by the user:\n'
-  for (const f of files) {
-    prefix += `\n## ${f.label}: ${f.path}\n`
-  }
-  return `${prefix}\n## My request for Codex:\n\n${prompt}\n`
-}
-
-function escapeHeartbeatXmlText(value: string): string {
-  return value
-    .replace(/&/gu, '&amp;')
-    .replace(/</gu, '&lt;')
-    .replace(/>/gu, '&gt;')
-}
-
-function buildHeartbeatQueuedMessage(automation: ThreadAutomationRecord): StoredQueuedMessage {
-  return {
-    id: `automation-${automation.id}-${Date.now()}-${randomBytes(3).toString('hex')}`,
-    text: `<heartbeat>
-<automation_id>${escapeHeartbeatXmlText(automation.id)}</automation_id>
-<current_time_iso>${new Date().toISOString()}</current_time_iso>
-<instructions>
-${escapeHeartbeatXmlText(automation.prompt)}
-</instructions>
-</heartbeat>`,
-    imageUrls: [],
-    skills: [],
-    fileAttachments: [],
-    collaborationMode: 'default',
-  }
-}
-
-function fileNameFromPath(pathValue: string): string {
-  const normalized = pathValue.replace(/\\/g, '/')
-  const segments = normalized.split('/').filter(Boolean)
-  return segments.at(-1) ?? normalized
-}
-
-function extractThreadIdFromNotificationParams(params: unknown): string {
-  const record = asRecord(params)
-  if (!record) return ''
-  const threadId =
-    (typeof record.threadId === 'string' ? record.threadId : '') ||
-    (typeof record.thread_id === 'string' ? record.thread_id : '') ||
-    (typeof record.conversationId === 'string' ? record.conversationId : '') ||
-    (typeof record.conversation_id === 'string' ? record.conversation_id : '')
-  if (threadId) return threadId
-  const thread = asRecord(record.thread)
-  if (thread && typeof thread.id === 'string') return thread.id
-  const turn = asRecord(record.turn)
-  if (turn) {
-    const turnThreadId =
-      (typeof turn.threadId === 'string' ? turn.threadId : '') ||
-      (typeof turn.thread_id === 'string' ? turn.thread_id : '')
-    if (turnThreadId) return turnThreadId
-  }
-  return ''
-}
-
-function isTurnCompletedNotification(notification: { method: string; params: unknown }): boolean {
-  return notification.method === 'turn/completed'
-}
-
 async function rollbackCreatedWorktree(
   gitRoot: string,
   worktreeCwd: string,
@@ -1883,838 +1655,6 @@ async function fetchConnectorLogo(rawUrl: string): Promise<{ contentType: string
   return {
     contentType: contentType || 'image/png',
     body: Buffer.from(await response.arrayBuffer()),
-  }
-}
-
-const STREAM_EVENT_BUFFER_LIMIT = 400
-
-type StreamEventFrame = {
-  method: string
-  params: unknown
-  atIso: string
-}
-
-type CapturedItem = {
-  id: string
-  type: string
-  turnId: string
-  data: Record<string, unknown>
-  completed: boolean
-}
-
-const MERGEABLE_ITEM_TYPES = new Set([
-  'commandExecution',
-  'fileChange',
-])
-
-class AppServerProcess {
-  private process: ChildProcessWithoutNullStreams | null = null
-  private initialized = false
-  private initializePromise: Promise<void> | null = null
-  private readBuffer = ''
-  private nextId = 1
-  private stopping = false
-  private readonly pending = new Map<number, { resolve: (value: unknown) => void; reject: (reason?: unknown) => void }>()
-  private readonly notificationListeners = new Set<(value: { method: string; params: unknown }) => void>()
-  private readonly pendingServerRequests = new Map<number, PendingServerRequest>()
-  private readonly appServerArgs = buildAppServerArgs()
-  private readonly streamEventsByThreadId = new Map<string, StreamEventFrame[]>()
-  private readonly lastThreadReadSnapshotByThreadId = new Map<string, unknown>()
-  private readonly threadTurnPageReadCacheByThreadId = new Map<string, { result: unknown; expiresAt: number }>()
-  private readonly threadTurnPageReadPromiseByThreadId = new Map<string, Promise<unknown>>()
-  private readonly capturedItemsByThreadId = new Map<string, Map<string, CapturedItem>>()
-  private readonly liveStateCache = new Map<string, { data: unknown; turnCount: number; sessionSize: number }>()
-  private chatgptAuthRefreshPromise: Promise<ChatgptAuthTokensRefreshResponse> | null = null
-
-
-  private getCodexCommand(): string {
-    const codexCommand = resolveCodexCommand()
-    if (!codexCommand) {
-      throw new Error('Codex CLI is not available. Install @openai/codex or set CODEXUI_CODEX_COMMAND.')
-    }
-    return codexCommand
-  }
-
-  private buildAppServerConfig(): { args: string[]; env: Record<string, string> } {
-    const args = [
-      'app-server',
-      '-c', 'approval_policy="never"',
-      '-c', 'sandbox_mode="danger-full-access"',
-    ]
-    let extraEnv: Record<string, string> = {}
-    const serverPort = parseInt(process.env.CODEXUI_SERVER_PORT ?? '', 10) || undefined
-    const statePath = join(getCodexHomeDir(), FREE_MODE_STATE_FILE)
-    try {
-      const state = ensureDefaultFreeModeStateForMissingAuthSync(statePath)
-      if (state) {
-        args.push(...getFreeModeConfigArgs(state, serverPort))
-        extraEnv = getFreeModeEnvVars(state)
-      }
-    } catch {
-      // No free-mode state or invalid — use defaults
-    }
-    return { args, env: extraEnv }
-  }
-
-  private start(): void {
-    if (this.process) return
-
-    this.stopping = false
-    const config = this.buildAppServerConfig()
-    const invocation = getSpawnInvocation(this.getCodexCommand(), config.args)
-    const spawnEnv = Object.keys(config.env).length > 0
-      ? { ...process.env, ...config.env }
-      : undefined
-    const proc = spawn(invocation.command, invocation.args, { stdio: ['pipe', 'pipe', 'pipe'], ...(spawnEnv ? { env: spawnEnv } : {}) })
-    this.process = proc
-
-    proc.stdout.setEncoding('utf8')
-    proc.stdout.on('data', (chunk: string) => {
-      this.readBuffer += chunk
-
-      let lineEnd = this.readBuffer.indexOf('\n')
-      while (lineEnd !== -1) {
-        const line = this.readBuffer.slice(0, lineEnd).trim()
-        this.readBuffer = this.readBuffer.slice(lineEnd + 1)
-
-        if (line.length > 0) {
-          this.handleLine(line)
-        }
-
-        lineEnd = this.readBuffer.indexOf('\n')
-      }
-    })
-
-    proc.stderr.setEncoding('utf8')
-    proc.stderr.on('data', () => {
-      // Keep stderr silent in dev middleware; JSON-RPC errors are forwarded via responses.
-    })
-
-    proc.on('exit', () => {
-      if (this.process !== proc) {
-        return
-      }
-
-      const failure = new Error(this.stopping ? 'codex app-server stopped' : 'codex app-server exited unexpectedly')
-      for (const request of this.pending.values()) {
-        request.reject(failure)
-      }
-
-      this.pending.clear()
-      this.pendingServerRequests.clear()
-      this.process = null
-      this.initialized = false
-      this.initializePromise = null
-      this.readBuffer = ''
-    })
-  }
-
-  private sendLine(payload: Record<string, unknown>): void {
-    if (!this.process) {
-      throw new Error('codex app-server is not running')
-    }
-
-    this.process.stdin.write(`${JSON.stringify(payload)}\n`)
-  }
-
-  private handleLine(line: string): void {
-    let message: JsonRpcResponse
-    try {
-      message = JSON.parse(line) as JsonRpcResponse
-    } catch {
-      return
-    }
-
-    if (typeof message.id === 'number' && this.pending.has(message.id)) {
-      const pendingRequest = this.pending.get(message.id)
-      this.pending.delete(message.id)
-
-      if (!pendingRequest) return
-
-      if (message.error) {
-        pendingRequest.reject(new Error(message.error.message))
-      } else {
-        pendingRequest.resolve(message.result)
-      }
-      return
-    }
-
-    if (typeof message.method === 'string' && typeof message.id !== 'number') {
-      this.emitNotification({
-        method: message.method,
-        params: message.params ?? null,
-      })
-      return
-    }
-
-    // Handle server-initiated JSON-RPC requests (approvals, dynamic tool calls, etc.).
-    if (typeof message.id === 'number' && typeof message.method === 'string') {
-      this.handleServerRequest(message.id, message.method, message.params ?? null)
-    }
-  }
-
-  private emitNotification(notification: { method: string; params: unknown }): void {
-    this.recordStreamEvent(notification)
-    this.captureItemFromNotification(notification)
-    const nThreadId = this.extractThreadIdFromParams(notification.params)
-    if (nThreadId) {
-      this.invalidateLiveStateCache(nThreadId)
-      this.threadTurnPageReadCacheByThreadId.delete(nThreadId)
-    }
-    for (const listener of this.notificationListeners) {
-      listener(notification)
-    }
-  }
-
-  private extractThreadIdFromParams(params: unknown): string {
-    const record = asRecord(params)
-    if (!record) return ''
-    const threadId =
-      (typeof record.threadId === 'string' ? record.threadId : '') ||
-      (typeof record.thread_id === 'string' ? record.thread_id : '') ||
-      (typeof record.conversationId === 'string' ? record.conversationId : '') ||
-      (typeof record.conversation_id === 'string' ? record.conversation_id : '')
-    if (threadId) return threadId
-    const thread = asRecord(record.thread)
-    if (thread && typeof thread.id === 'string') return thread.id
-    const turn = asRecord(record.turn)
-    if (turn) {
-      const turnThreadId =
-        (typeof turn.threadId === 'string' ? turn.threadId : '') ||
-        (typeof turn.thread_id === 'string' ? turn.thread_id : '')
-      if (turnThreadId) return turnThreadId
-    }
-    return ''
-  }
-
-  private recordStreamEvent(notification: { method: string; params: unknown }): void {
-    const threadId = this.extractThreadIdFromParams(notification.params)
-    if (!threadId) return
-    const frame: StreamEventFrame = {
-      method: notification.method,
-      params: notification.params,
-      atIso: new Date().toISOString(),
-    }
-    let buffer = this.streamEventsByThreadId.get(threadId)
-    if (!buffer) {
-      buffer = []
-      this.streamEventsByThreadId.set(threadId, buffer)
-    }
-    buffer.push(frame)
-    if (buffer.length > STREAM_EVENT_BUFFER_LIMIT) {
-      buffer.splice(0, buffer.length - STREAM_EVENT_BUFFER_LIMIT)
-    }
-  }
-
-  getStreamEvents(threadId: string, limit: number): StreamEventFrame[] {
-    const buffer = this.streamEventsByThreadId.get(threadId)
-    if (!buffer || buffer.length === 0) return []
-    return buffer.slice(-limit)
-  }
-
-  storeThreadReadSnapshot(threadId: string, snapshot: unknown): void {
-    this.lastThreadReadSnapshotByThreadId.set(threadId, snapshot)
-    this.threadTurnPageReadCacheByThreadId.delete(threadId)
-  }
-
-  getLastThreadReadSnapshot(threadId: string): unknown | null {
-    return this.lastThreadReadSnapshotByThreadId.get(threadId) ?? null
-  }
-
-  async readThreadForTurnPage(threadId: string): Promise<unknown> {
-    const now = Date.now()
-    const cached = this.threadTurnPageReadCacheByThreadId.get(threadId)
-    if (cached && cached.expiresAt > now) return cached.result
-    if (cached) this.threadTurnPageReadCacheByThreadId.delete(threadId)
-
-    const pending = this.threadTurnPageReadPromiseByThreadId.get(threadId)
-    if (pending) return pending
-
-    const promise = this.rpc('thread/read', {
-      threadId,
-      includeTurns: true,
-    }).then((result) => {
-      this.threadTurnPageReadCacheByThreadId.set(threadId, {
-        result,
-        expiresAt: Date.now() + THREAD_TURN_PAGE_READ_CACHE_TTL_MS,
-      })
-      return result
-    }).finally(() => {
-      this.threadTurnPageReadPromiseByThreadId.delete(threadId)
-    })
-
-    this.threadTurnPageReadPromiseByThreadId.set(threadId, promise)
-    return promise
-  }
-
-  cacheLiveState(threadId: string, data: unknown, turnCount: number, sessionSize: number): void {
-    this.liveStateCache.set(threadId, { data, turnCount, sessionSize })
-  }
-
-  getCachedLiveState(threadId: string, turnCount: number, sessionSize: number): unknown | null {
-    const cached = this.liveStateCache.get(threadId)
-    if (!cached) return null
-    if (cached.turnCount !== turnCount || cached.sessionSize !== sessionSize) return null
-    return cached.data
-  }
-
-  invalidateLiveStateCache(threadId: string): void {
-    this.liveStateCache.delete(threadId)
-  }
-
-  private captureItemFromNotification(notification: { method: string; params: unknown }): void {
-    if (notification.method !== 'item/started' && notification.method !== 'item/completed') return
-
-    const params = asRecord(notification.params)
-    if (!params) return
-    const item = asRecord(params.item)
-    if (!item) return
-    const itemType = typeof item.type === 'string' ? item.type : ''
-    if (!MERGEABLE_ITEM_TYPES.has(itemType)) return
-
-    const itemId = typeof item.id === 'string' ? item.id : ''
-    if (!itemId) return
-
-    const threadId = this.extractThreadIdFromParams(params)
-    if (!threadId) return
-
-    const turnId =
-      (typeof params.turnId === 'string' ? params.turnId : '') ||
-      (typeof params.turn_id === 'string' ? params.turn_id : '')
-    if (!turnId) return
-
-    let threadItems = this.capturedItemsByThreadId.get(threadId)
-    if (!threadItems) {
-      threadItems = new Map()
-      this.capturedItemsByThreadId.set(threadId, threadItems)
-    }
-
-    const isCompleted = notification.method === 'item/completed'
-    const existing = threadItems.get(itemId)
-
-    if (existing && existing.completed && !isCompleted) return
-
-    threadItems.set(itemId, {
-      id: itemId,
-      type: itemType,
-      turnId,
-      data: item as Record<string, unknown>,
-      completed: isCompleted,
-    })
-  }
-
-  mergeItemsIntoTurns(threadId: string, turns: unknown[]): unknown[] {
-    const capturedMap = this.capturedItemsByThreadId.get(threadId)
-    if (!capturedMap || capturedMap.size === 0) return turns
-
-    const itemsByTurnId = new Map<string, CapturedItem[]>()
-    for (const captured of capturedMap.values()) {
-      let group = itemsByTurnId.get(captured.turnId)
-      if (!group) {
-        group = []
-        itemsByTurnId.set(captured.turnId, group)
-      }
-      group.push(captured)
-    }
-
-    return turns.map((turn) => {
-      const turnRecord = asRecord(turn)
-      if (!turnRecord) return turn
-      const turnId = typeof turnRecord.id === 'string' ? turnRecord.id : ''
-      if (!turnId) return turn
-
-      const captured = itemsByTurnId.get(turnId)
-      if (!captured || captured.length === 0) return turn
-
-      const existingItems = Array.isArray(turnRecord.items) ? (turnRecord.items as Record<string, unknown>[]) : []
-      const existingIds = new Set(existingItems.map((it) => (typeof it.id === 'string' ? it.id : '')).filter(Boolean))
-
-      const newItems = captured
-        .filter((c) => !existingIds.has(c.id))
-        .map((c) => c.data)
-
-      if (newItems.length === 0) return turn
-
-      return {
-        ...turnRecord,
-        items: [...existingItems, ...newItems],
-      }
-    })
-  }
-
-  private sendServerRequestReply(requestId: number, reply: ServerRequestReply): void {
-    if (reply.error) {
-      this.sendLine({
-        jsonrpc: '2.0',
-        id: requestId,
-        error: reply.error,
-      })
-      return
-    }
-
-    this.sendLine({
-      jsonrpc: '2.0',
-      id: requestId,
-      result: reply.result ?? {},
-    })
-  }
-
-  private resolvePendingServerRequest(requestId: number, reply: ServerRequestReply): void {
-    const pendingRequest = this.pendingServerRequests.get(requestId)
-    if (!pendingRequest) {
-      throw new Error(`No pending server request found for id ${String(requestId)}`)
-    }
-    this.pendingServerRequests.delete(requestId)
-
-    this.sendServerRequestReply(requestId, reply)
-    const requestParams = asRecord(pendingRequest.params)
-    const threadId =
-      typeof requestParams?.threadId === 'string' && requestParams.threadId.length > 0
-        ? requestParams.threadId
-        : ''
-    this.emitNotification({
-      method: 'server/request/resolved',
-      params: {
-        id: requestId,
-        method: pendingRequest.method,
-        threadId,
-        mode: 'manual',
-        resolvedAtIso: new Date().toISOString(),
-      },
-    })
-  }
-
-  private async refreshChatgptAuthTokens(params: ChatgptAuthTokensRefreshParams): Promise<ChatgptAuthTokensRefreshResponse> {
-    if (!this.chatgptAuthRefreshPromise) {
-      this.chatgptAuthRefreshPromise = refreshChatgptAuthTokensForExternalAuth(params).finally(() => {
-        this.chatgptAuthRefreshPromise = null
-      })
-    }
-    return await this.chatgptAuthRefreshPromise
-  }
-
-  private async handleChatgptAuthTokensRefreshRequest(requestId: number, params: unknown): Promise<void> {
-    const requestParams = asRecord(params)
-    const previousAccountId = readNonEmptyString(requestParams?.previousAccountId ?? requestParams?.previous_account_id)
-    try {
-      const result = await this.refreshChatgptAuthTokens({
-        reason: readNonEmptyString(requestParams?.reason) || undefined,
-        previousAccountId: previousAccountId || undefined,
-      })
-      this.sendServerRequestReply(requestId, { result })
-      this.emitNotification({
-        method: 'server/request/resolved',
-        params: {
-          id: requestId,
-          method: 'account/chatgptAuthTokens/refresh',
-          mode: 'automatic',
-          resolvedAtIso: new Date().toISOString(),
-        },
-      })
-    } catch (error) {
-      this.sendServerRequestReply(requestId, {
-        error: {
-          code: -32001,
-          message: getErrorMessage(error, 'Failed to refresh ChatGPT auth tokens'),
-        },
-      })
-    }
-  }
-
-  private handleServerRequest(requestId: number, method: string, params: unknown): void {
-    if (method === 'account/chatgptAuthTokens/refresh') {
-      void this.handleChatgptAuthTokensRefreshRequest(requestId, params)
-      return
-    }
-
-    const pendingRequest: PendingServerRequest = {
-      id: requestId,
-      method,
-      params,
-      receivedAtIso: new Date().toISOString(),
-    }
-    this.pendingServerRequests.set(requestId, pendingRequest)
-
-    this.emitNotification({
-      method: 'server/request',
-      params: pendingRequest,
-    })
-  }
-
-  private async call(method: string, params: unknown): Promise<unknown> {
-    this.start()
-    const id = this.nextId++
-
-    return new Promise((resolve, reject) => {
-      this.pending.set(id, { resolve, reject })
-
-      this.sendLine({
-        jsonrpc: '2.0',
-        id,
-        method,
-        params,
-      } satisfies JsonRpcCall)
-    })
-  }
-
-  private async ensureInitialized(): Promise<void> {
-    if (this.initialized) return
-    if (this.initializePromise) {
-      await this.initializePromise
-      return
-    }
-
-    this.initializePromise = this.call('initialize', {
-      clientInfo: {
-        name: 'codex-web-local',
-        version: '0.1.0',
-      },
-      capabilities: {
-        experimentalApi: true,
-      },
-    }).then(() => {
-      this.sendLine({
-        jsonrpc: '2.0',
-        method: 'initialized',
-      })
-      this.initialized = true
-    }).finally(() => {
-      this.initializePromise = null
-    })
-
-    await this.initializePromise
-  }
-
-  async rpc(method: string, params: unknown): Promise<unknown> {
-    await this.ensureInitialized()
-    return this.call(method, params)
-  }
-
-  onNotification(listener: (value: { method: string; params: unknown }) => void): () => void {
-    this.notificationListeners.add(listener)
-    return () => {
-      this.notificationListeners.delete(listener)
-    }
-  }
-
-  async respondToServerRequest(payload: unknown): Promise<void> {
-    await this.ensureInitialized()
-
-    const body = asRecord(payload)
-    if (!body) {
-      throw new Error('Invalid response payload: expected object')
-    }
-
-    const id = body.id
-    if (typeof id !== 'number' || !Number.isInteger(id)) {
-      throw new Error('Invalid response payload: "id" must be an integer')
-    }
-
-    const rawError = asRecord(body.error)
-    if (rawError) {
-      const message = typeof rawError.message === 'string' && rawError.message.trim().length > 0
-        ? rawError.message.trim()
-        : 'Server request rejected by client'
-      const code = typeof rawError.code === 'number' && Number.isFinite(rawError.code)
-        ? Math.trunc(rawError.code)
-        : -32000
-      this.resolvePendingServerRequest(id, { error: { code, message } })
-      return
-    }
-
-    if (!('result' in body)) {
-      throw new Error('Invalid response payload: expected "result" or "error"')
-    }
-
-    this.resolvePendingServerRequest(id, { result: body.result })
-  }
-
-  listPendingServerRequests(): PendingServerRequest[] {
-    return Array.from(this.pendingServerRequests.values())
-  }
-
-  dispose(): void {
-    if (!this.process) return
-
-    const proc = this.process
-    this.stopping = true
-    this.process = null
-    this.initialized = false
-    this.initializePromise = null
-    this.readBuffer = ''
-
-    const failure = new Error('codex app-server stopped')
-    for (const request of this.pending.values()) {
-      request.reject(failure)
-    }
-    this.pending.clear()
-    this.pendingServerRequests.clear()
-
-    try {
-      proc.stdin.end()
-    } catch {
-      // ignore close errors on shutdown
-    }
-
-    try {
-      proc.kill('SIGTERM')
-    } catch {
-      // ignore kill errors on shutdown
-    }
-
-    const forceKillTimer = setTimeout(() => {
-      if (!proc.killed) {
-        try {
-          proc.kill('SIGKILL')
-        } catch {
-          // ignore kill errors on shutdown
-        }
-      }
-    }, 1500)
-    forceKillTimer.unref()
-  }
-}
-
-export class BackendQueueProcessor {
-  private readonly processingThreadIds = new Set<string>()
-  private readonly queueDrainTimersByThreadId = new Map<string, ReturnType<typeof setTimeout>>()
-  private readonly queueDrainDueAtByThreadId = new Map<string, number>()
-  private readonly unsubscribe: () => void
-
-  constructor(private readonly appServer: AppServerProcess) {
-    this.unsubscribe = appServer.onNotification((notification) => {
-      if (!isTurnCompletedNotification(notification)) return
-      const threadId = extractThreadIdFromNotificationParams(notification.params)
-      if (!threadId) return
-      void this.processThreadQueue(threadId)
-    })
-    void this.scheduleAllQueuedThreads(1000)
-  }
-
-  dispose(): void {
-    this.unsubscribe()
-    for (const timer of this.queueDrainTimersByThreadId.values()) {
-      clearTimeout(timer)
-    }
-    this.queueDrainTimersByThreadId.clear()
-    this.queueDrainDueAtByThreadId.clear()
-    this.processingThreadIds.clear()
-  }
-
-  async scheduleAllQueuedThreads(delayMs = 0): Promise<void> {
-    try {
-      const state = await readThreadQueueState()
-      for (const threadId of Object.keys(state)) {
-        this.scheduleThreadQueueDrain(threadId, delayMs)
-      }
-    } catch {
-      // Queue recovery is best-effort; normal turn-completed events can still drain later.
-    }
-  }
-
-  scheduleThreadQueueDrain(threadId: string, delayMs = 5000): void {
-    if (!threadId) return
-    const normalizedDelayMs = Math.max(0, delayMs)
-    const nextDueAt = Date.now() + normalizedDelayMs
-    const existingDueAt = this.queueDrainDueAtByThreadId.get(threadId)
-    const existingTimer = this.queueDrainTimersByThreadId.get(threadId)
-    if (existingTimer) {
-      if (existingDueAt !== undefined && existingDueAt <= nextDueAt) return
-      clearTimeout(existingTimer)
-      this.queueDrainTimersByThreadId.delete(threadId)
-      this.queueDrainDueAtByThreadId.delete(threadId)
-    }
-    const timer = setTimeout(() => {
-      this.queueDrainTimersByThreadId.delete(threadId)
-      this.queueDrainDueAtByThreadId.delete(threadId)
-      void this.processThreadQueue(threadId)
-    }, normalizedDelayMs)
-    timer.unref?.()
-    this.queueDrainTimersByThreadId.set(threadId, timer)
-    this.queueDrainDueAtByThreadId.set(threadId, nextDueAt)
-  }
-
-  async processThreadQueue(threadId: string): Promise<void> {
-    if (this.processingThreadIds.has(threadId)) return
-    this.processingThreadIds.add(threadId)
-    try {
-      const canStart = await this.canStartQueuedTurn(threadId)
-      if (!canStart) {
-        if (await this.hasQueuedTurns(threadId)) {
-          this.scheduleThreadQueueDrain(threadId)
-        }
-        return
-      }
-      const next = await this.popNextQueuedTurn(threadId)
-      if (!next) return
-      try {
-        await this.startQueuedTurn(next)
-        if (await this.hasQueuedTurns(threadId)) {
-          this.scheduleThreadQueueDrain(threadId)
-        }
-      } catch {
-        await this.restoreQueuedTurn(next)
-        this.scheduleThreadQueueDrain(threadId)
-      }
-    } catch {
-      // Queue processing is best-effort. Keep the bridge alive if app-server is unavailable.
-      this.scheduleThreadQueueDrain(threadId)
-    } finally {
-      this.processingThreadIds.delete(threadId)
-    }
-  }
-
-  private async hasQueuedTurns(threadId: string): Promise<boolean> {
-    const state = await readThreadQueueState()
-    const queue = state[threadId]
-    return Array.isArray(queue) && queue.length > 0
-  }
-
-  private async canStartQueuedTurn(threadId: string): Promise<boolean> {
-    const response = asRecord(await this.appServer.rpc('thread/read', { threadId, includeTurns: true }))
-    const thread = asRecord(response?.thread)
-    if (!thread) return false
-
-    const status = asRecord(thread.status)
-    const statusType = readNonEmptyString(status?.type)
-    if (statusType === 'inProgress' || statusType === 'running' || statusType === 'active') return false
-
-    const turns = Array.isArray(thread.turns) ? thread.turns : []
-    return !turns.some((turn) => readNonEmptyString(asRecord(turn)?.status) === 'inProgress')
-  }
-
-  private async popNextQueuedTurn(threadId: string): Promise<BackendQueuedTurn | null> {
-    return withThreadQueueStateUpdate((state) => {
-      const queue = state[threadId]
-      if (!queue || queue.length === 0) {
-        return { nextState: state, result: null }
-      }
-
-      const [message, ...rest] = queue
-      const nextState = { ...state }
-      if (rest.length > 0) {
-        nextState[threadId] = rest
-      } else {
-        delete nextState[threadId]
-      }
-      return { nextState, result: { threadId, message } }
-    })
-  }
-
-  private async restoreQueuedTurn(turn: BackendQueuedTurn): Promise<void> {
-    await withThreadQueueStateUpdate((state) => {
-      const queue = state[turn.threadId] ?? []
-      return {
-        nextState: {
-          ...state,
-          [turn.threadId]: [turn.message, ...queue],
-        },
-        result: undefined,
-      }
-    })
-  }
-
-  private async resolveCollaborationModeSettings(mode: CollaborationModeKind): Promise<ResolvedCollaborationModeSettings> {
-    let currentConfig: Record<string, unknown> | null = null
-    try {
-      const configPayload = asRecord(await this.appServer.rpc('config/read', {}))
-      currentConfig = asRecord(configPayload?.config)
-    } catch {
-      currentConfig = null
-    }
-
-    const configuredModel = readNonEmptyString(currentConfig?.model)
-    if (configuredModel) {
-      return {
-        model: configuredModel,
-        reasoningEffort: normalizeCollaborationModeReasoningEffort(normalizeReasoningEffort(currentConfig?.model_reasoning_effort)),
-      }
-    }
-
-    try {
-      const modelsPayload = asRecord(await this.appServer.rpc('model/list', {}))
-      const models = Array.isArray(modelsPayload?.data) ? modelsPayload.data : []
-      for (const row of models) {
-        const record = asRecord(row)
-        const candidate = readNonEmptyString(record?.id) || readNonEmptyString(record?.model)
-        if (candidate) {
-          return {
-            model: candidate,
-            reasoningEffort: normalizeCollaborationModeReasoningEffort(normalizeReasoningEffort(currentConfig?.model_reasoning_effort)),
-          }
-        }
-      }
-    } catch {
-      // Fall through to no collaboration-mode payload.
-    }
-
-    throw new Error(`${mode === 'plan' ? 'Plan' : 'Default'} mode requires an available model.`)
-  }
-
-  private async buildQueuedTurnParams(turn: BackendQueuedTurn): Promise<Record<string, unknown>> {
-    const localImageAttachments: StoredQueuedMessage['fileAttachments'] = []
-    for (const imageUrl of turn.message.imageUrls) {
-      const localImagePath = extractLocalImagePathFromUrl(imageUrl.trim())
-      if (!localImagePath) continue
-      localImageAttachments.push({
-        label: fileNameFromPath(localImagePath),
-        path: localImagePath,
-        fsPath: localImagePath,
-      })
-    }
-
-    const allFileAttachments = [...turn.message.fileAttachments, ...localImageAttachments]
-    const dedupedFileAttachments = allFileAttachments.filter((entry, index) =>
-      allFileAttachments.findIndex((candidate) => candidate.fsPath === entry.fsPath) === index)
-
-    const input: Array<Record<string, unknown>> = [{
-      type: 'text',
-      text: buildTextWithAttachments(turn.message.text, dedupedFileAttachments),
-    }]
-
-    for (const imageUrl of turn.message.imageUrls) {
-      const normalizedUrl = imageUrl.trim()
-      if (!normalizedUrl) continue
-      const localImagePath = extractLocalImagePathFromUrl(normalizedUrl)
-      if (localImagePath) {
-        input.push({ type: 'localImage', path: localImagePath })
-      } else {
-        input.push({ type: 'image', url: normalizedUrl, image_url: normalizedUrl })
-      }
-    }
-
-    for (const skill of turn.message.skills) {
-      input.push({ type: 'skill', name: skill.name, path: skill.path })
-    }
-
-    const params: Record<string, unknown> = {
-      threadId: turn.threadId,
-      input,
-    }
-    if (dedupedFileAttachments.length > 0) {
-      params.attachments = dedupedFileAttachments.map((f) => ({ label: f.label, path: f.path, fsPath: f.fsPath }))
-    }
-
-    try {
-      const settings = await this.resolveCollaborationModeSettings(turn.message.collaborationMode)
-      params.collaborationMode = {
-        mode: turn.message.collaborationMode,
-        settings: {
-          model: settings.model,
-          reasoning_effort: settings.reasoningEffort,
-          developer_instructions: null,
-        },
-      }
-    } catch {
-      // Older app-server versions still accept a plain turn/start without collaborationMode.
-    }
-
-    return params
-  }
-
-  private async startQueuedTurn(turn: BackendQueuedTurn): Promise<void> {
-    await this.appServer.rpc('thread/resume', { threadId: turn.threadId })
-    await this.appServer.rpc('turn/start', await this.buildQueuedTurnParams(turn))
   }
 }
